add second example
